requirements.txt
Django==3.2.16
python-docx